requirements.txt
firebase-admin==4.0.1
Flask==1.1.1
Flask-Cors==3.0.8
Flask-RESTful==0.3.8
FormEncode==1.3.1
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.13.1
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-firestore==1.6.2
google-cloud-logging==1.15.0
google-cloud-storage==1.27.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
gunicorn==20.0.4
grpcio==1.28.1
python-dateutil==2.8.1
requests==2.23.0
pycryptodomex==3.9.7
Sphinx==3.1.2
rinohtype==0.4.1
